requirements.txt
Flask==2.1.0
flask-kafka
kafka-python
sumolib
